Refactor packages names for better encapsulation
