+++ packages/sanes-chrome-extension/src/extension/bsExternalHandler/senderWhitelist.ts
